+++ source/extensions/filters/network/mongo_proxy/config.cc
@@ -32,12 +32,13 @@ Network::FilterFactoryCb MongoProxyFilterConfigFactory::createFilterFactoryFromP
     fault_config = std::make_shared<Filters::Common::Fault::FaultDelayConfig>(proto_config.delay());
   }
 
+  auto stats = std::make_shared<MongoStats>(context.scope(), stat_prefix);
   const bool emit_dynamic_metadata = proto_config.emit_dynamic_metadata();
-  return [stat_prefix, &context, access_log, fault_config,
-          emit_dynamic_metadata](Network::FilterManager& filter_manager) -> void {
+  return [stat_prefix, &context, access_log, fault_config, emit_dynamic_metadata,
+          stats](Network::FilterManager& filter_manager) -> void {
     filter_manager.addFilter(std::make_shared<ProdProxyFilter>(
         stat_prefix, context.scope(), context.runtime(), access_log, fault_config,
-        context.drainDecision(), context.dispatcher().timeSource(), emit_dynamic_metadata));
+        context.drainDecision(), context.dispatcher().timeSource(), emit_dynamic_metadata, stats));
   };
 }